incorp test files into main files to hopefull merge easier
